Consolidate "event" and "gesture" layers into a single layer by using manager.emit(). Alias basic and gestural events as necessary. Add documentation and cleanup.
