+++ src/utils/event-manager.js
@@ -1,137 +1,216 @@
-import {createGestureManager, EVENT_RECOGNIZER_MAP} from './gesture-manager';
+import {
+  Manager,
+  Tap,
+  Press,
+  Pinch,
+  Rotate,
+  Pan,
+  Swipe
+} from 'hammerjs';
+import WheelInput from './wheel-input';
 
-const EVENT_HANDLER_MAP = {
-  pointerdown: 'mousedown',
-  pointermove: 'mousemove',
-  pointerup: 'mouseup',
-  mousedown: 'mousedown',
-  mousemove: 'mousemove',
-  mouseup: 'mouseup'
+/**
+ * Only one set of basic input events will be fired by Hammer.js:
+ * either pointer, touch, or mouse, depending on system support.
+ * In order to enable an application to be agnostic of system support,
+ * alias basic input events into "classes" of events: down, move, and up.
+ * See `_onBasicInput()` for usage of these aliases.
+ */
+/* eslint-disable */
+const BASIC_EVENT_CLASSES = {
+  down: ['pointerdown', 'touchstart', 'mousedown'],
+  move: ['pointermove', 'touchmove',  'mousemove'],
+  up:   ['pointerup',   'touchend',   'mouseup']
 };
-const GESTURE_ALIASES = {
-  click: 'tap'
-};
+/* eslint-enable */
 
-const DEFAULT_OPTIONS = {
-  handleRawEventsWithGestureManager: true
+const BASIC_EVENT_ALIASES = {
+  pointerdown: BASIC_EVENT_CLASSES.down,
+  pointermove: BASIC_EVENT_CLASSES.move,
+  pointerup: BASIC_EVENT_CLASSES.up,
+  touchstart: BASIC_EVENT_CLASSES.down,
+  touchmove: BASIC_EVENT_CLASSES.move,
+  touchend: BASIC_EVENT_CLASSES.up,
+  mousedown: BASIC_EVENT_CLASSES.down,
+  mousemove: BASIC_EVENT_CLASSES.move,
+  mouseup: BASIC_EVENT_CLASSES.up
 };
 
 /**
- * Single API for subscribing to events about both "raw" input (e.g. 'mousemove', 'click')
- * and gestural input (e.g. 'panstart', 'tap').
+ * "Gestural" events are those that have semantic meaning beyond the basic input event,
+ * e.g. a click or tap is a sequence of `down` and `up` events with no `move` event in between.
+ * Hammer.js handles these with its Recognizer system;
+ * this block maps event names to the Recognizers required to detect the events.
+ */
+const EVENT_RECOGNIZER_MAP = {
+  click: 'tap',
+  tap: 'tap',
+  doubletap: 'tap',
+  press: 'press',
+  pinch: 'pinch',
+  pinchin: 'pinch',
+  pinchout: 'pinch',
+  pinchstart: 'pinch',
+  pinchmove: 'pinch',
+  pinchend: 'pinch',
+  pinchcancel: 'pinch',
+  rotate: 'rotate',
+  rotatestart: 'rotate',
+  rotatemove: 'rotate',
+  rotateend: 'rotate',
+  rotatecancel: 'rotate',
+  pan: 'pan',
+  panstart: 'pan',
+  panmove: 'pan',
+  panup: 'pan',
+  pandown: 'pan',
+  panleft: 'pan',
+  panright: 'pan',
+  panend: 'pan',
+  pancancel: 'pan',
+  swipe: 'swipe',
+  swipeleft: 'swipe',
+  swiperight: 'swipe',
+  swipeup: 'swipe',
+  swipedown: 'swipe'
+};
+
+const RECOGNIZERS = {
+  doubletap: new Tap({
+    event: 'doubletap',
+    taps: 2
+  }),
+  pan: new Pan({
+    threshold: 10
+  }),
+  pinch: new Pinch(),
+  press: new Press(),
+  rotate: new Rotate(),
+  swipe: new Swipe(),
+  tap: new Tap()
+};
+
+/**
+ * Map gestural events typically provided by browsers
+ * that are not reported in 'hammer.input' events
+ * to corresponding Hammer.js gestures.
+ */
+const GESTURE_EVENT_ALIASES = {
+  tap: 'click'
+};
+
+/**
+ * Single API for subscribing to events about both
+ * basic input events (e.g. 'mousemove', 'touchstart')
+ * and gestural input (e.g. 'click', 'tap', 'panstart').
+ * Delegates event registration and handling to Hammer.js.
  */
 class EventManager {
   constructor(element, options) {
-    this.element = element;
-    this.handlers = {};
-    this.gestureManager = createGestureManager(element, options);
+    // TODO: support overriding default RECOGNIZERS by passing
+    // recognizers / configs, keyed to event name.
 
-    // If specified, use GestureManager for raw event handling
-    // as well as for gesture recognition.
-    if (options.handleRawEventsWithGestureManager) {
-      this._onRawInput = this._onRawInput.bind(this);
-      this.gestureManager.startHandlingRawEvents(this._onRawInput);
-    }
+    // how to get inputClass from createInputInstance without a Manager instance?
+    // mostly just need to run logic from createInputInstance...
+    // Issue filed: https://github.com/hammerjs/hammer.js/issues/1106
+    const inputClass = WheelInput;
+
+    this._onBasicInput = this._onBasicInput.bind(this);
+    this.manager = new Manager(element, {
+      inputClass
+    })
+    .on('hammer.input', this._onBasicInput);
+
+    this.gestureAliases = {};
   }
 
+  /**
+   * Register an event handler function to be called on `event`.
+   * @param {string|Object} event   An event name (String) or map of event names to handlers
+   * @param {Function} [handler]    The function to be called on `event`.
+   */
   on(event, handler) {
     if (typeof event === 'string') {
-      // wrap and store handler
-      if (!this.handlers[event]) {
-        this.handlers[event] = [];
-      }
-      const handlersForEvent = this.handlers[event];
-      let handlerForEvent = handlersForEvent.find(h => h.handler === handler);
-      if (!handlerForEvent) {
-        handlerForEvent = {
-          handler,
-          wrapper: e => handler(this._wrapEvent(e, 'srcEvent'))
-        };
-        handlersForEvent.push(handlerForEvent);
+      // Special handling for gestural events.
+      const recognizerEvent = EVENT_RECOGNIZER_MAP[event];
+      if (recognizerEvent) {
+        // Add recognizer for this event if not already added.
+        if (!this.manager.get(recognizerEvent)) {
+          this.manager.add(RECOGNIZERS[recognizerEvent]);
+        }
+
+        // Alias to a recognized gesture as necessary.
+        const gestureEventAlias = GESTURE_EVENT_ALIASES[recognizerEvent];
+        if (gestureEventAlias) {
+          this.gestureAliases[event] = this._wrapAliasedGestureHandler(gestureEventAlias, handler);
+          this.manager.on(gestureEventAlias, this.gestureAliases[event]);
+        }
       }
 
-      // Handle as a gestural event if appropriate
-      const aliasedEvent = GESTURE_ALIASES[event] || event;
-      if (EVENT_RECOGNIZER_MAP[aliasedEvent]) {
-        this.gestureManager.on(aliasedEvent, handlerForEvent.wrapper);
-      }
+      // Register event handler.
+      this.manager.on(recognizerEvent || event, handler);
     } else {
+      // If `event` is a map, call `on()` for each entry.
       for (const eventName in event) {
         this.on(eventName, event[eventName]);
       }
     }
   }
 
+  /**
+   * Deregister a previously-registered event handler.
+   * @param {string|Object} event   An event name (String) or map of event names to handlers
+   * @param {Function} [handler]    The function to be called on `event`.
+   */
   off(event, handler) {
     if (typeof event === 'string') {
-      // find wrapped handler
-      const handlersForEvent = this.handlers[event];
-      if (handlersForEvent) {
-        const i = handlersForEvent.findIndex(h => h.handler === handler);
-        let wrapper;
-        if (i !== -1) {
-          wrapper = handlersForEvent.splice(i, 1)[0].wrapper;
-        }
-        if (!handlersForEvent.length) {
-          delete this.handlers[event];
-        }
-
-        // Handle as a gestural event if appropriate
-        const aliasedEvent = GESTURE_ALIASES[event] || event;
-        if (wrapper && EVENT_RECOGNIZER_MAP[aliasedEvent]) {
-          this.gestureManager.off(aliasedEvent, wrapper);
+      // Clean up aliased gesture handler as necessary.
+      const recognizerEvent = EVENT_RECOGNIZER_MAP[event];
+      if (recognizerEvent) {
+        const gestureEventAlias = GESTURE_EVENT_ALIASES[recognizerEvent];
+        if (gestureEventAlias) {
+          this.manager.off(gestureEventAlias, this.gestureAliases[event]);
+          delete this.gestureAliases[event];
         }
       }
+
+      // Deregister event handler.
+      this.manager.off(event, handler);
     } else {
+      // If `event` is a map, call `off()` for each entry.
       for (const eventName in event) {
         this.off(eventName, event[eventName]);
       }
     }
   }
 
-  _onRawInput(event) {
-    if (EVENT_RECOGNIZER_MAP[event.type]) {
-      // let GestureManager handle these events
-      return;
-    }
-
+  /**
+   * Handle basic events using the 'hammer.input' Hammer.js API:
+   * Before running Recognizers, Hammer emits a 'hammer.input' event
+   * with the basic event info. This function emits all basic events
+   * aliased to the "class" of event received.
+   * See BASIC_EVENT_CLASSES basic event class definitions.
+   */
+  _onBasicInput(event) {
     const {srcEvent} = event;
-    const normalizedEventType = srcEvent && EVENT_HANDLER_MAP[srcEvent.type];
-    if (!normalizedEventType) {
-      // not a recognized event type
-      return;
-    }
-
-    const handlersForEvent = this.handlers[normalizedEventType];
-    if (handlersForEvent) {
-      handlersForEvent.forEach(handler => handler.wrapper(event));
+    const eventAliases = BASIC_EVENT_ALIASES[srcEvent.type];
+    if (eventAliases) {
+      // fire all events aliased to srcEvent.type
+      eventAliases.forEach(alias => {
+        const emitEvent = Object.assign({}, event, {type: alias});
+        this.manager.emit(alias, emitEvent);
+      });
     }
   }
 
-  _wrapEvent(event, srcEventPropName) {
-    // TODO: decide on how best to wrap for cross-browser compatibility
-    // possible options:
-    // - luma.gl's `EventsProxy.eventInfo()` wrapper
-    //    (https://github.com/uber/luma.gl/blob/master/src/addons/event.js#L171)
-    // - npm `synthetic-dom-event`
-    //    (https://www.npmjs.com/package/synthetic-dom-events))
-    // - hammer.js `compute-input-data` wrapper + additional info
-    //    (https://github.com/hammerjs/hammer.js/blob/master/src/inputjs/compute-input-data.js)
-
-    // TODO: consider refactoring deck.gl to look for `srcEvent` instead of `event`
-    // for original/source event object;
-    // `event.srcEvent` is more descriptive than `event.event`
-
-    return {
-      event: event[srcEventPropName]
-    };
+  _wrapAliasedGestureHandler(eventAlias) {
+    return event => this.manager.emit(eventAlias, event);
   }
 }
 
 export function createEventManager(element, options = {}) {
   // TODO: should this be a Singleton?
-  // conversely, is actually a good reason to use the Factory pattern here,
-  // or should we allow direct instantiation?
-  options = Object.assign({}, DEFAULT_OPTIONS, options);
+  // conversely, is there actually a good reason to use the Factory pattern here,
+  // or should we just allow direct instantiation?
   return new EventManager(element, options);
 }
